Relocated files into main folder and subfolders
