+++ src/components/page/main/ProjectsSection/ProjectsSections.js
@@ -1,7 +1,7 @@
 import * as React from 'react'
 
-import { Banner } from './Banner'
-import Carousel from './Carousel'
+import { Banner } from '../shared/Banner'
+import Carousel from '../shared/Carousel'
 import ProjectTemplate from './ProjectTemplate'
 import projects from 'content/projects'
